Applying Ryan's patch for moving back ColumnFormatting
