+++ engine/table/src/main/java/io/deephaven/engine/table/impl/by/AggAllByExcludeFormattingColumns.java
@@ -1,7 +1,7 @@
 /**
  * Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
  */
-package io.deephaven.qst.table;
+package io.deephaven.engine.table.impl.by;
 
 import io.deephaven.api.agg.spec.*;
