+++ engine/table/src/test/java/io/deephaven/engine/util/TestColorUtil.java
@@ -8,7 +8,6 @@ import io.deephaven.engine.context.TestExecutionContext;
 import io.deephaven.engine.table.impl.DataAccessHelpers;
 import io.deephaven.gui.color.Color;
 import io.deephaven.engine.table.Table;
-import io.deephaven.qst.table.ColumnFormatting;
 import io.deephaven.util.QueryConstants;
 import io.deephaven.util.SafeCloseable;
